feat: wrap up colletoral score
