+++ src/component-library/CollateralScore/index.tsx
@@ -1,0 +1,2 @@
+export type { CollateralScoreProps } from './CollateralScore';
+export { CollateralScore } from './CollateralScore';
