--- src/component-library/ScoreBar/index.tsx
@@ -1,2 +1,0 @@
-export type { ScoreBarProps } from './ScoreBar';
-export { ScoreBar } from './ScoreBar';
